+++ .github/generate-workflows.py
@@ -46,7 +46,7 @@ for documentclass in documentclasses:
                 continue
               if (documentclass == 'ieee'):
                 dashedPart = "{}-{}-{}-{}-{}-{}-{}".format(documentclass, ieeevariant, papersize, latexcompiler, bibtextool, texlive, example);
-                dashedPartMiktex = "{}-{}-{}-{}-{}-{}".format(documentclass, ieeevariant, papersize, latexcompiler, bibtextool, example);
+                dashedPartMiktex = "{}-{}-{}-{}-{}-{}".format(documentclass, ieeevariant, papersize, latexcompiler, bibtextool, example, newline='\n');
               else:
                 dashedPart = "{}-{}-{}-{}-{}-{}".format(documentclass, papersize, latexcompiler, bibtextool, texlive, example);
                 dashedPartMiktex = "{}-{}-{}-{}-{}".format(documentclass, papersize, latexcompiler, bibtextool, example);
@@ -54,6 +54,8 @@ for documentclass in documentclasses:
               yml.write("name: Check {}\n".format(dashedPart))
               yml.write("""on:
   push:
+    branches:
+      - main
     paths-ignore:
       - '.editorconfig'
       - '.eslintignore'
@@ -71,12 +73,35 @@ for documentclass in documentclasses:
       - 'user-data.sh'
       - 'docs/**'
       - '.vscode/**'
+  pull_request:
+    branches:
+      - main
+    paths-ignore:
+      - '.editorconfig'
+      - '.eslintignore'
+      - '.gitpod.dockerfile'
+      - '.gitpod.yml'
+      - '.gitattributes'
+      - '.gitignore'
+      - '.markdownlint.yml'
+      - 'CHANGELOG.md'
+      - 'CONTRIBUTING.md'
+      - 'generate-texlivefile.sh'
+      - 'LICENSE'
+      - 'README.md'
+      - 'setup-do.sh'
+      - 'user-data.sh'
+      - 'docs/**'
+      - '.vscode/**'
+  merge_group:
 concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
+  group: "${{ github.workflow }}-${{ github.head_ref || github.ref }}"
   cancel-in-progress: true
 jobs:
   check:
-    runs-on: ubuntu-latest
+""")
+              yml.write("    name: Check {}\n".format(dashedPart))
+              yml.write("""    runs-on: ubuntu-latest
     services:
       registry:
         image: registry:2
@@ -84,11 +109,11 @@ jobs:
           - 5000:5000
     steps:
       - name: Set up Git repository
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
       - name: Set up QEMU
-        uses: docker/setup-qemu-action@v2
+        uses: docker/setup-qemu-action@v3
       - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v2
+        uses: docker/setup-buildx-action@v3
         with:
           driver-opts: network=host
       - name: Cache Docker layers
@@ -96,9 +121,9 @@ jobs:
         with:
           path: /tmp/.buildx-cache
           key: ${{ runner.os }}-buildx
-      - uses: actions/setup-node@v3
+      - uses: actions/setup-node@v4
         with:
-          node-version: '14'
+          node-version: '20'
       - name: Update npm
         run: |
           npm i -g npm@latest
@@ -114,6 +139,7 @@ jobs:
               ymlmiktex.write("  cancel-in-progress: true\n")
               ymlmiktex.write("jobs:\n")
               ymlmiktex.write("  miktex:\n")
+              ymlmiktex.write("    name: MiKTeX {}\n".format(dashedPartMiktex))
               ymlmiktex.write('''    runs-on: windows-latest
     steps:
       - name: Install MikTeX
@@ -132,17 +158,10 @@ jobs:
           initexmf --admin --verbose --mklinks --force
           updmap --admin
       - name: Checkout repository
-        uses: actions/checkout@v3
-      - uses: actions/setup-node@v3
-        with:
-          node-version: '14'
-      - name: Update npm
-        run: |
-          npm i -g npm@latest
-          npm i npm@latest
-      - run: npm install
+        uses: actions/checkout@v4
 ''')
-              table = "documentclass | latexcompiler | bibtextool | texlive | lang | font    | listing  | enquote    | tweakouterquote | todo       | example | howtotext\n"
+              table = "| documentclass | latexcompiler | bibtextool | texlive | lang | font    | listing  | enquote    | tweakouterquote | todo       | example | howtotext |\n"
+              table += "| -- | -- | -- | -- | -- | --| -- | -- | -- | -- | -- | -- |\n"
               for howtotext in howtotexts:
                 for language in languages:
                   for font in fonts:
@@ -153,12 +172,12 @@ jobs:
                         for tweakouterquote in tweakouterquotes:
                           for todo in todos:
                               variantName = "{}_{}_{}_{}_{}_{}_{}_{}_{}_{}_{}_{}".format(documentclass, latexcompiler, bibtextool, texlive, language, font, listing, enquote, tweakouterquote, todo, example, howtotext)
-                              table += "{:<13} | {:<13} | {:<10} | {:<7} | {:<4} | {:<7} | {:<8} | {:10} | {:<15} | {:<10} | {:<7} | {:<8}\n".format(documentclass, latexcompiler, bibtextool, texlive, language, font, listing, enquote, tweakouterquote, todo, example, howtotext)
+                              table += "| {:<13} | {:<13} | {:<10} | {:<7} | {:<4} | {:<7} | {:<8} | {:10} | {:<15} | {:<10} | {:<7} | {:<8} |\n".format(documentclass, latexcompiler, bibtextool, texlive, language, font, listing, enquote, tweakouterquote, todo, example, howtotext)
                               yml_content = "      - run: mkdir {}\n".format(variantName)
                               yml_content += "      - name: Create {}\n".format(variantName)
                               yml_content += '''        shell: bash
         run: |
-          npx yo $GITHUB_WORKSPACE\\
+          npx yo@v4.3.1 $GITHUB_WORKSPACE\\
 '''
                               yml_content += "           --documentclass=%s\\\n" % documentclass
                               if documentclass == 'ieee':
@@ -187,10 +206,12 @@ jobs:
                               yml.write(yml_content)
                               ymlmiktex.write(yml_content)
                               yml.write('''      - name: Build docker image
-        uses: docker/build-push-action@v4
+        uses: docker/build-push-action@v5
         with:
           push: true
           provenance: false
+          build-args: |
+            TLMIRRORURL=https://tug.ctan.org/
 ''')
                               yml.write("          tags: localhost:5000/name/app:{}\n".format(variantName))
                               yml.write("          context: '${{{{ github.workspace }}}}/{}'\n".format(variantName))
@@ -204,7 +225,7 @@ jobs:
                               ymlmiktex.write("        run: {}\n".format(command))
                               yml.write("        working-directory: '${{{{ github.workspace }}}}/{}'\n".format(variantName))
                               ymlmiktex.write("        working-directory: '${{{{ github.workspace }}}}/{}'\n".format(variantName))
-              yml.write('''      - uses: actions/upload-artifact@v2
+              yml.write('''      - uses: actions/upload-artifact@v4
         with:
           name: pdfs
           path: /tmp/out